+++ Mikado/utilities/intervaltree.pyx
@@ -253,6 +253,7 @@ cdef class IntervalNode:
         cdef list results = []
         # use end + 1 becuase .right() assumes strictly right-of
         self._seek_right(position + 1, results, n, max_dist)
+
         if len(results) == n: return results
         r = results
         r.sort(key=operator.attrgetter('start'))
@@ -307,19 +308,19 @@ cdef class Interval:
     def __richcmp__(self, other, op):
         if op == 0:
             # <
-            return self.start < other.start or self.end < other.end
+            return self[0] < other[0] or self[1] < other[1]
         elif op == 1:
             # <=
             return self == other or self < other
         elif op == 2:
             # ==
-            return self.start == other.start and self.end == other.end
+            return self[0] == other[0] and self[1] == other[1]
         elif op == 3:
             # !=
-            return self.start != other.start or self.end != other.end
+            return self[0] != other[0] or self[1] != other[1]
         elif op == 4:
             # >
-            return self.start > other.start or self.end > other.end
+            return self[0] > other[0] or self[1] > other[1]
         elif op == 5:
             # >=
             return self == other or self > other
@@ -329,8 +330,16 @@ cdef class Interval:
             return self.start
         elif index == 1:
             return self.end
+        elif index == 2:
+            return self.value
         else:
-            raise IndexError("Intervals only have starts and ends!")
+            return [self.start, self.end, self.value][index]
+            # raise IndexError("Intervals only have starts and ends!")
+
+    def __iter__(self):
+
+        return iter([self.start, self.end])
+
 
 cdef class IntervalTree:
     """
@@ -406,9 +415,11 @@ cdef class IntervalTree:
 
     add = insert
 
-    def find(self, int start, int end, bint strict=0, int max_distance=0, int num_intervals=100):
+    def find(self, int start, int end, bint strict=0, int max_distance=0, int num_intervals=100, object value=None):
         """
         Return a sorted list of all intervals overlapping [start,end).
+        If strict is set to True, only matches which are completely contained will
+        be counted as full matches. If set to False, also partial matches will count.
         """
 
         if self.root is None:
@@ -430,6 +441,13 @@ cdef class IntervalTree:
                     new_found.append(_)
             found = new_found
 
+        if value is not None:
+            new_found = []
+            for _ in found:
+               if _.value == value:
+                   new_found.append(_)
+            found = new_found
+
         return found
 
     search = find
@@ -439,23 +457,23 @@ cdef class IntervalTree:
 
         return self.num_intervals
 
-    def before( self, position, num_intervals=1, max_dist=2000 ):
+    def before( self, position, num_intervals=1, max_dist=2000):
         """
         Find `num_intervals` intervals that lie before `position` and are no
         further than `max_dist` positions away
         """
         if self.root is None:
             return []
-        return self.root.left( position, num_intervals, max_dist )
+        return self.root.left( position, num_intervals, max_dist)
 
-    def after( self, position, num_intervals=1, max_dist=25000 ):
+    def after( self, position, num_intervals=1, max_dist=25000):
         """
         Find `num_intervals` intervals that lie after `position` and are no
         further than `max_dist` positions away
         """
         if self.root is None:
             return []
-        return self.root.right( position, num_intervals, max_dist )
+        return self.root.right( position, num_intervals, max_dist)
 
     # ---- Interval-like object based interfaces -----------------------------
 
@@ -530,6 +548,18 @@ cdef class IntervalTree:
         for iv in tuples:
             tree.insert_interval(Interval(*iv))
         return tree
+
+    @classmethod
+    def from_intervals(cls, intervals):
+        """
+        Create a new IntervalTree from an iterable of Interval instances.
+        """
+
+        tree = IntervalTree()
+        for iv in intervals:
+            tree.insert_interval(iv)
+        return tree
+
 
 # For backward compatibility
 Intersecter = IntervalTree